Use local entry costs as heuristics for the A-star path finding algorithm
Use the entry costs times the distance to target to estimate the future costs for a node in the A-star path finding algorithm. The original heuristic was the minimum road costs within a four times four square. This biased ship routes along costs with roads, railroads, or maglevs, even so the direct route over the ocean would be shorter.

The other problem with this heuristic is that it is not necessarily in the same scale as the distance, the trade A-star uses trade values that may be just half the value and the city A-star uses the build costs of the tile improvements or some measure that depends on it.

However, the new heuristics is still not perfect since it still underestimates or overestimates the future costs if some mountains are on the path, so it still may give sub-optimal paths. But that is a property of the heuristic and fine as long as we are not too far away from the optimum.
..\ctp2_code\robot\pathing\Astar.h

Improve A-star debugging if you define PRINT_COSTS in astar.h and you have enabled /debugcells via the chat window, you will get the total costs printed onto the tiles while you are pathing.
..\ctp2_code\robot\pathing\Astar.h
..\ctp2_code\robot\pathing\astar.cpp
..\ctp2_code\gfx\gfx_utils\gfx_options.cpp
..\ctp2_code\gfx\gfx_utils\gfx_options.h

While adding the A-star debugging feature, I also fixed a memory leak.
..\ctp2_code\gfx\gfx_utils\gfx_options.cpp
..\ctp2_code\gfx\gfx_utils\gfx_options.h