The AI now builds missionaries (if in theology) and diplomats
The AI now builds missionaries if it is in theology. That is a bit questionable since the televangelists are not really available when the AI is in theology. Another problem is that the AI is not long in theology even so if it is preferred government since all the AI personalities use almost the same advance order for  research.

The AI now also builds diplomats, but it still does not establish embassies.

..\ctp2_code\ai\CityManagement\governor.cpp
..\ctp2_code\ai\CityManagement\governor.h
..\ctp2_code\gs\newdb\strategy.cdb
..\ctp2_data\default\aidata\strategies.txt
..\ctp2_data\default\aidata\unitbuildlists.txt