Make the AI using Capitalization and Infrastructure
The AI adds Capitalization and Infrastructure to its build queues.
..\ctp2_code\ai\CityManagement\governor.cpp

Simplified handling of Capitalization and Infrastructure. A city is now Capitalization and Infrastructure if those are in the build queue instead of using Boolean variables. That was needed, so that Capitalization and Infrastructure can be removed when something else was added to the build queue after Capitalization or Infrastructure. Actually, this way Capitalization and Infrastructure did not give their boni to the player. This was actually a bug.

Now also Capitalization and Infrastructure are removed from the build queue, if they are followed by another item in the build queue.
..\ctp2_code\gs\gameobj\CityData.cpp
..\ctp2_code\gs\gameobj\citydata.h

Made GetHead const and some cleaning
..\ctp2_code\gs\gameobj\BldQue.h
..\ctp2_code\gs\gameobj\bldque.cpp