Make the AI only consider goals it has armies for
The AI now only adds goals to the scheduler for consideration if it has armies that can execute them. This saves processor time and memory.
..\ctp2_code\ai\ctpai.cpp
..\ctp2_code\ai\strategy\scheduler\Scheduler.h
..\ctp2_code\ai\strategy\scheduler\scheduler.cpp

Cities now build their units before the AI adds any goals to the scheduler. The change above would not have considered the units in the build queues finishing this turn.
..\ctp2_code\gs\gameobj\PlayerTurn.cpp

Added some annotations
..\ctp2_code\gs\gameobj\Unit.cpp
..\ctp2_code\gs\utility\TurnCntEvent.cpp