Installations are removed when other Tile imps remove their tile imps
Installations are removed if their tile improvement is removed, because another improvement was built there. This removes a bug where the AI build Gaia towers/obelisks and put afterwards another tile improvement on top. The obelisk tile improvement is removed but it is still listed as installation, and thus still shows up in the Gaia Controller map.
..\ctp2_code\gs\gameobj\TerrImproveData.cpp