Fix the settle map, the AI now selects settle target from the whole map
Now the AI selects settle targets from the whole map. Before it selected settle targets from a bounding rect. However, the bounding rect was not properly computed for x-wraps, so that the bounding was too smal it did not have the empire in its core. Of course eventually, the AI should also settle all the available spots.
..\ctp2_code\ai\ctpai.cpp
..\ctp2_code\ai\mapanalysis\settlemap.cpp

Removed the bounding rect calculation from the map analysis, since it is not needed anymore.
..\ctp2_code\ai\mapanalysis\mapanalysis.cpp
..\ctp2_code\ai\mapanalysis\mapanalysis.h

Marked the bounding rect files for later removal.
..\ctp2_code\ai\mapanalysis\boundingrect.cpp
..\ctp2_code\ai\mapanalysis\boundingrect.h