Splitter civs inherit the age of their parent civ
So far when a city becomes independent and forms a new civilization it inherits all the advances of its parent civilization. Since the advances define, in which age a civilization is, the age should also be inherited. Otherwise the new civilization is the first age until it discovers a new advance.
..\ctp2_code\gs\gameobj\Civilisation.cpp