Accelerate debugging when looking what units an AI has
Do not sort the list of the unit manager while rebuilding it, that is just a waste of time. This is useful if your AI has built some 800 units.
..\ctp2_code\ui\interface\unitmanager.cpp

Do not sort the entries in a list box while you are removing all of them. This actually applies to the list box of the unit manager, but of course it should also apply to all other lists boxes in the game.
..\ctp2_code\ui\aui_ctp2\ctp2_listbox.cpp

While at it, remove some unnecessary new lines, this is the only change to this file.
..\ctp2_code\ui\aui_common\aui_listbox.cpp