Clerics and Televangelists can now be build under any government.
For now, this is something for testing. This is a way to solve the problem that the AI can only build Clerics and Televangelists for a very short time, since it is only for roughly a few turns in Theocracy the time needed to research Fascism.

There are still alternatives to slove this problem for instance to introduce Fascim later in the tech tree.

..\ctp2_data\default\gamedata\Units.txt