Don't show city capture options for cities with 1 pop
When capturing a city with population of 1, it is automatically razed yet
the City Capture Options dialog is still shown. The dialog buttons can be
clicked but it has no effect, since city is already destroyed.

Note: the observed behavior is with AE Mod scernario. I haven't played vanilla
for decades, so I'm not entirely sure that cities lose population after being
captured in vanilla game.