[linux] Partial fix for permanent 100% CPU usage
The event loop throttling for 30 msec implemented by this path reduced the game
CPU load on my laptop (Thinkpad X220 i7 2.7 Ghz) from 100% to around 50% in menu
and 55% while in-game. Not ideal solution, but it prevents the CPU cooler noise
from overheat at 100%.

The proper solution would be to rework the event loop to the blocking mode with
SDL_WaitEvent(). Yet the current hack is still better than nothing.

fixes #272