Prevent tooltips from stealing keyboard events
Tooltpis are implemented as a special type of a window `AUI_WINDOW_TYPE_TIP`. So
while a tooltip is shown, it becomes a current focused window that gets to
handle all keyboard input, and this in turn prevents an actual useful
window (e.g. Build Manager) from handling keyboard.

This commit solves the issue by closing a tooltip before handling keyboard
input.