Fix comparing an unsigned int to -2
This caused the compiler to complain. There seems to have onvce been code that sorted the data in a default way when -2 was passed to the sort routine, but this seems to have been removed.

A search of the code did not turn up any cases where this is still used. ctp_ListBox and c3_ListBox will now return an error if -2 is passed as the column in SortByColumn.