Fix crash when removing the first goal
The code can remove goals from the list, but would crash if the first goal was removed as it would try to set the iterator to the previous goal, then remove the goal - which doesn't work if the goal to be removed is the first goal.

This is rather more elegant and with far less code change thanks to review by William Deurwaarder.