Ensure same looping sound can be added again shortly after termination
When a looping sound is terminated, it may still be available in the list. In case the same sound was added again, it was not restarted as it was found in the list. The condition has been improved so that the sound in the list will be re-used and played.