Technical: fix memory leak (sort of) while saving game
During saving the game an extra UnseenInstallationInfo-record was added to every unseen cell to store the city owner. This record was not removed.

This caused additional memory to be allocated after every save game (cumulative). (Re-)loading the game removed this additional memory. As the record was stored in a list it was correctly deleted when quitting the game.

This was mostly noticeable due to auto-save and playing for a long time without reloading or quitting the game.

A side-effect was that the save-game size increased as well (cumulative). This was also resetted when the game was reloaded and saved again.