Fix decrements where goal_iter could be become invalid
There are places in the code where goal_iter is decremented, then reincremented due to this initially being implemented as a for loop. This does not work is the goal_iter is the first item in the list, as goal_iter willl become invalid when decremented.

Now that the loop has been changes from a for loop to a while loop, these can be fixed.