Fix decrement where agent_iter could be become invalid
There is a place in the code where agent_iter is decremented, then reincremented by the for loop - this does not work is the goal_iter is the first item in the list, as agent_iter willl become invalid when decremented.

Change the loop to a while loop and replace the decrement with a continue to fix this.