Remove unused GetPaletteIndexColor routines
The RGBQUAD version of GetPaletteIndexColor can cause a compilation error because it compares an uint8 to 256, which is more then this type can hold. The HPALETTE version won't actually work, as it casts a value of 256 to a unit8 (which will give a result of zero) - therefore the function will always return 0.

A search of the code shows that these functions are not used anywhere and a search of the git history indicates this has been the case since the original commit, therefore this removes the functions entirely.