Align city icons and information; fog is applied to all icons
City icons and information is aligned. Capitol, sleeping unit, and watchfull are shown next to the city name. In case the city is in 'fog' all icons are darkened completely. The isRioting icon has been deprecated as the name of the city is already indicating this state (red). All frames are made a single pixel in the color of the player.

Some new icons are introduced (updated) to be able to use the color of the player as frame.