Introduce extra zoom-level to improve drawing quality
Not all tile heights (in pixels) of the different zoom-levels were a multiple of 4 which caused all kind of rounding errors while calculating pixel positions. By splitting two zoom levels, which were 6 pixels apart, to three levels that are 4 pixels apart all tile heights (in pixels) are a multiple of 4. As tile widths (in pixels) are twice the tile heights (in pixels) they are also a multiple of 4.