Fix issues introduced to AI scheduler by #311
In two cases goal_iter would not be pointing to the correct item in the m_goals list in the next loop iteration:

1.    When a goal is erased
2.    When Reprioritize_Goal moves the current goal to later in the list.

This fixes the first issue by setting goal_iter to the iterator return by erase() and the second issue by changing Reprioritize_Goal to follow the same pattern and return the corrrect iterator to use.

This also fixes the debugging loopCount to only increment when moving on to the next item in the list - not erasing, replacing or repeating the current item.