Ensure that all edges can be reached on all zoom levels
Sometimes with a certain zoom-level the edge of the map was not shown and units could not be selected. Now left and top are aligned with the first tile, number of black tiles on the right depends on the screen size, and number of black tiles at the bottom is aligned with the height of the control panel.