Technical: Include right and bottom pixels in rectangle draw methods
Primitives paint rectangle methods, such as primitives_ClippedPaintRect16 or primitives_ClippedFrameRect16 did not include the right and bottom pixel, which is confusing. This is fixed and two new methods (primitives_ClippedPaint16 and primitives_ClippedFrame16) are introduced that have as input: x, y, width, and height. These methods can be used in case the right and bottom should be excluded.