Clamp RelativeY to 0 when negative
Fixes crash caused by clicking on the top border of an empty listbox.

RelativeY is negative in this case, which bypasses sanity checks which
assume a non-negative value. This becomes a negative index, which
causes a segfault later upon attempted deference.

Might be nice if some of the assumptions on paramters in this class were
more explicit.