use magic number from orig code for quick trade distance calculation:
in the orig code CalcTerrainFreightCost just returned 5 as a constant: https://github.com/civctp2/civctp2/blob/d614fbdf705db334a4b45037bbcc735142d22016/ctp2_code/gs/world/wldgen.cpp#L1984-L1990
in c0adde0314, c9f40abe59  CalcTerrainFreightCost was properly implemented
but the new implementation should only be used for TradeAstar but not for the quick calculation
because just the value CalcTerrainFreightCost for a single tile (the source pos of a good)
is not appropriate as a factor for the normalized distance to estimating a quick trade distance
e.g. if the source pos is on a glacier tile the whole trade distance is scale up a lot even if all other tiles of the trade route would be e.g. water