Reduce the amount of siege goals the AI tries to execute
If the AI has not enough units for the top siege goals it would go down the list for cities that are less optimal to siege but need fewer units and are further away. However if in the next turns the AI gets more units, it would then abort these sieges, wasting time. So better use these units for defense until the AI has produced enough units.