Revert "Radarmap: Toggle Unit & Political view so only one is selected"
This reverts commit 19b8067d8d56b828bf26e48213a0eade4a3535b6.

Actually, this only reverts radarwindow.cpp and reimplements the dark color selection in radamap.cpp, since radamap.cpp has been heavily refactored.

The original reason for the commit is bogus, the map is simply not less cluttered with outline borders, add terrain to it and it is more cluttered. Actually, whether to use this feature is up to the player.