Add ARM and ARM 64 platform to the ctp2 project, remove DebugNoMem configuration
ARM and ARM 64 seem to compile, but do not link due to missing libraries. The are no SDL2 builds for ARM and ARM64. SDL3 include ARM64 builds but there is no SDL3 mixer, yet. The alternative would be to compile from source.

The FFmpeg has some ARM64 prebuild libraries, but the zip seems to be broken to include the dll files. Unfortunately the source contains assembly code that can be ported to different platforms but it seems only x86 and x64 platforms are supported.

Even dinput8.lib seems not be available, so that DirectX builds do not work either. at least there is now quite some preparation in the project for ARM64 support.