Add the exported script and images files from tile edit
The script has the added border tiles and for some reason it a few altered tile rules and the images seem to differ by some odd pixels. Of course neither the tile rules nor these images now included should have been changed. However, that does not seem to be a problem for the tile files themselves.