Fix x64 VS2017 warnings
This are the last warnings in the ctp2 project VS2017 reports that do not come from automatically generated files. For those files Byacc and Flex would be needed to be updated or a newer version of them.

VS2017 does not report warnings for x84 SDL builds and two warnings for DirectX related code in DirectX x84 builds.