Make gcc compile the code again
Looks like VS2017 loves to compile gcc doesn't like to compile

This uses nullptr a feature of C++11. Obviously, that is not compatible with VS6, but that is long gone anyway. And there is already code in that uses newer features of C++. Probably all instances of the NULL macro should be replaced by nullptr, but be careful the code has also "Null" as string literal.