Replace atoi and atof by strtol and strtod in database parsing
This allows handling under and over flows in database handling. This way, the maxmum or minimum possible values are used on all platforms. The implementation of atoi on VS seems to do this by default. But it should be on all platforms so that MP games can be joined by all players whether they are on Windows or Linux.